requirements.txt
Django~=5.0.7
pillow~=10.4.0
rules~=3.4
factory-boy~=3.3.0
whitenoise~=6.7.0
gunicorn~=23.0.0
pylint~=3.2.6
pylint-django~=2.5.5
channels~=4.1.0
asgiref~=3.8.1
uvicorn~=0.30.6
websockets~=12.0
channels-redis~=4.2.0
celery~=5.5.0b2
redis~=5.0.8
openai~=1.42.0
pytest~=8.3.2
pytest-django~=4.8.0
django-bootstrap5~=24.2
djangorestframework~=3.15.2